+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Changelog
 Next
 ----
 
+2019.06.07.0
+------------
+
 * Get DC/OS build information from a ``Node`` object after installation.
 * Added option to ``minidcos aws create`` and ``minidcos aws provision`` to choose the AWS instance type.
 * Changed the default version of Docker installed on ``minidcos docker`` clusters to ``18.06.3-ce``.